Reorganize image factories and fix Recipe image widths
