+++ app/Services/Image/Factories/ImageFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Services\Image;
+namespace App\Services\Image\Factories;
 
 use App\Http\Traits\VariableClassName;
 
@@ -8,7 +8,7 @@ class ImageFactory
 {
     use VariableClassName;
 
-    const IMAGE_NAMESPACE = 'App\Services\Image';
+    const IMAGE_NAMESPACE = 'App\Services\Image\Factories';
 
     public function get(string $imageType)
     {
