+++ app/Services/Interfaces/Image/ImageProcessorInterface.php
@@ -3,7 +3,7 @@
 namespace App\Services\Interfaces\Image;
 
 use App\Services\File\SaverFactory;
-use App\Services\Image\ImageFactory;
+use App\Services\Image\Factories\ImageFactory;
 use Illuminate\Http\UploadedFile;
 use Intervention\Image\ImageManager;
